Replace expandable row collapsable with tabs - Add Tabs.tsx component - Closes #225
Update expandable row tooltip title
